Rename HTMLLocalization to Localization
