+++ src/runtime/tooling/index.js
@@ -3,12 +3,12 @@ import FTLEntriesParser from '../../ftl/entries/parser';
 import { createEntriesFromAST } from '../../ftl/entries/transformer';
 import { MessageContext } from '../../intl/context';
 
-import { Localization } from '../../lib/localization';
+import * as format from '../../lib/format';
 import { L10nError } from '../../lib/errors';
 
 import { fetchResource } from '../web/io';
 
 export default {
   FTLASTParser, FTLEntriesParser, createEntriesFromAST, MessageContext,
-  Localization, L10nError, fetchResource
+  format, L10nError, fetchResource
 };
