+++ src/runtime/web/api.js
@@ -1,2 +1,1 @@
-export { Localization } from '../../lib/localization';
 export { fetchResource } from './io';
